requirements.txt
fastapi==0.83.0
hypercorn==0.14.3
Pillow
python-multipart
